// contracts/typescript/models/MemberProfile.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HomeOfficeReservation API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface MemberProfile
 */
export interface MemberProfile {
    /**
     *
     */
    active: boolean;
    /**
     *
     */
    displayName: string;
    /**
     *
     */
    email: string;
    /**
     *
     */
    isAccountAdministrator: boolean;
    /**
     *
     */
    isEmployee: boolean;
    /**
     *
     */
    isManager: boolean;
    /**
     *
     */
    memberId: string;
    /**
     *
     */
    organizationId: string;
    /**
     *
     */
    organizationName: string;
}

/**
 * Check if a given object implements the MemberProfile interface.
 */
export function instanceOfMemberProfile(value: object): value is MemberProfile {
    if (!('active' in value) || value['active'] === undefined) return false;
    if (!('displayName' in value) || value['displayName'] === undefined) return false;
    if (!('email' in value) || value['email'] === undefined) return false;
    if (!('isAccountAdministrator' in value) || value['isAccountAdministrator'] === undefined) return false;
    if (!('isEmployee' in value) || value['isEmployee'] === undefined) return false;
    if (!('isManager' in value) || value['isManager'] === undefined) return false;
    if (!('memberId' in value) || value['memberId'] === undefined) return false;
    if (!('organizationId' in value) || value['organizationId'] === undefined) return false;
    if (!('organizationName' in value) || value['organizationName'] === undefined) return false;
    return true;
}

export function MemberProfileFromJSON(json: any): MemberProfile {
    return MemberProfileFromJSONTyped(json, false);
}

export function MemberProfileFromJSONTyped(json: any, ignoreDiscriminator: boolean): MemberProfile {
    if (json == null) {
        return json;
    }
    return {

        'active': json['active'],
        'displayName': json['displayName'],
        'email': json['email'],
        'isAccountAdministrator': json['isAccountAdministrator'],
        'isEmployee': json['isEmployee'],
        'isManager': json['isManager'],
        'memberId': json['memberId'],
        'organizationId': json['organizationId'],
        'organizationName': json['organizationName'],
    };
}

export function MemberProfileToJSON(json: any): MemberProfile {
    return MemberProfileToJSONTyped(json, false);
}

export function MemberProfileToJSONTyped(value?: MemberProfile | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'active': value['active'],
        'displayName': value['displayName'],
        'email': value['email'],
        'isAccountAdministrator': value['isAccountAdministrator'],
        'isEmployee': value['isEmployee'],
        'isManager': value['isManager'],
        'memberId': value['memberId'],
        'organizationId': value['organizationId'],
        'organizationName': value['organizationName'],
    };
}